rewrote and broke out tests to simplify
